requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
DAWG-Python==0.7.2
docopt==0.6.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flickrapi==2.4.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.0.0
oauthlib==3.1.0
Pillow==8.0.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rutermextract==0.3
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.2
